requirements.txt
SoundFile
sounddevice
munch
pydub
pyyaml
librosa
nltk
matplotlib
accelerate
transformers
einops
einops-exts
tqdm
typing
typing-extensions
pyaudio
phonemizer
wget
omegaconf
pytorch_lightning
git+https://github.com/resemble-ai/monotonic_align.git